Moves Lint checks to Mods.Common
