+++ OpenRA.Mods.Common/Lint/CheckActorReferences.cs
@@ -12,7 +12,7 @@ using System;
 using System.Reflection;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Lint
 {
 	public class CheckActorReferences : ILintPass
 	{
