+++ OpenRA.Mods.Common/Lint/CheckSequences.cs
@@ -13,7 +13,7 @@ using System.Linq;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Lint
 {
 	class CheckSequences : ILintPass
 	{
