+++ OpenRA.Mods.Common/Lint/LintBuildablePrerequisites.cs
@@ -11,10 +11,9 @@
 using System;
 using System.Linq;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Lint
 {
 	class LintBuildablePrerequisites : ILintPass
 	{
